django_backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
openai==1.13.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
srt==3.5.3
tqdm==4.66.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
urllib3==2.2.1
vosk==0.3.45
webrtcvad==2.0.10
websockets==12.0
zope.interface==6.2
